+++ mix/test_backend.ex
@@ -3,8 +3,7 @@ defmodule MyApp.Cldr do
 
   use Cldr,
     default_locale: "en",
-    # locales: ["en", "zh", "it", "ja", "zh-Hant", "fr", "de", "th", "id"],
-    locales: ["en", "zh", "zh-Hant", "it", "fr", "de", "th", "id"],
+    locales: ["en", "zh", "zh-Hant", "it", "fr", "de", "th", "id", "ru", "he", "pl"],
     precompile_transliterations: [{:latn, :arab}, {:arab, :thai}, {:arab, :latn}],
     providers: [Cldr.Number]