+++ CHANGELOG.md
@@ -1,3 +1,11 @@
+# Changelog for Cldr_Numbers v2.4.0
+
+This is the changelog for Cldr v2.4.0 released on _____, 2019.  For older changelogs please consult the release tag on [GitHub](https://github.com/kipcole9/cldr_numbers/tags)
+
+## Enhancements
+
+* Adds `Cldr.Number.Format.default_grouping_for/2` to return the default grouping of digits for a locale. This is useful for external number formats like [ex_cldr_print](https://github.com/kipcole9/cldr_print).
+
 # Changelog for Cldr_Numbers v2.3.0
 
 This is the changelog for Cldr v2.3.0 released on March 1st, 2019.  For older changelogs please consult the release tag on [GitHub](https://github.com/kipcole9/cldr_numbers/tags)